Add numbering part; elements; properties
This is a start at trying to implement enough of the "numbering" part to
allow bulleted lists and numbered lists to be properly formatted.
Numbered lists turn out to be massively overcomplicated in
WordprocessingML and require a complete section dedicated to describing
them and which paragraphs belong to which numbered section.